Prevent Endpoint leakage when infra container is not running.
While using HNS V1, we do not fail the DEL request on failure of endpoint detachment if the infra container is not running.
While using HNS V2, even if the endpoint detachment from the namespace fails, we try to delete the endpoint. This is because there is 1-1 relation of endpoint to HCN namespace as opposed to 1-many relation while using HNS V1.